+++ services/horizon/internal/simplepath/finder.go
@@ -61,7 +61,6 @@ func (f *Finder) Find(q paths.Query, maxLength uint) (result []paths.Path, err e
 // FindFixedPaths will return an error because this implementation
 // does not support this operation
 func (f *Finder) FindFixedPaths(
-	sourceAccount *xdr.AccountId,
 	sourceAsset xdr.Asset,
 	amountToSpend xdr.Int64,
 	destinationAssets []xdr.Asset,