+++ services/horizon/internal/simplepath/inmemory.go
@@ -51,6 +51,7 @@ func (finder InMemoryFinder) Find(q paths.Query, maxLength uint) ([]paths.Path, 
 		q.SourceAccount,
 		q.SourceAssets,
 		q.SourceAssetBalances,
+		q.ValidateSourceBalance,
 		maxAssetsPerPath,
 	)
 	results := make([]paths.Path, len(orderbookPaths))
@@ -72,7 +73,6 @@ func (finder InMemoryFinder) Find(q paths.Query, maxLength uint) ([]paths.Path, 
 // `sourceAccountID` is optional. if `sourceAccountID` is provided then no offers
 // created by `sourceAccountID` will be considered when evaluating payment paths
 func (finder InMemoryFinder) FindFixedPaths(
-	sourceAccount *xdr.AccountId,
 	sourceAsset xdr.Asset,
 	amountToSpend xdr.Int64,
 	destinationAssets []xdr.Asset,
@@ -91,7 +91,6 @@ func (finder InMemoryFinder) FindFixedPaths(
 
 	orderbookPaths, err := finder.graph.FindFixedPaths(
 		int(maxLength),
-		sourceAccount,
 		sourceAsset,
 		amountToSpend,
 		destinationAssets,